btrfs: reada: Ignore creating reada_extent for non-exist device
For a non-exist device, old code bypass adding it in dev's reada
queue.

And to solve problem of unfinished waitting in raid5/6,
commit 5fbc7c59fd22 ("Btrfs: fix unfinished readahead thread for
raid5/6 degraded mounting")
adding a exception for the first strip, in short, the first
strip will always be processed whether the device exist or not.

Actually we have a better way for above request: just bypass
creation of the reada_extent for non-exist device, it will make
code simple and effective.

Signed-off-by: Zhao Lei <zhaolei@cn.fujitsu.com>